+++ app/src/main/java/com/forrestguice/suntimeswidget/LightMapDialog.java
@@ -1,5 +1,5 @@
 /**
-    Copyright (C) 2017 Forrest Guice
+    Copyright (C) 2017-2018 Forrest Guice
     This file is part of SuntimesWidget.
 
     SuntimesWidget is free software: you can redistribute it and/or modify
@@ -29,17 +29,27 @@ import android.support.annotation.NonNull;
 import android.support.v4.app.DialogFragment;
 import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AlertDialog;
+import android.text.SpannableString;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.ImageView;
+import android.widget.TextView;
 
 import com.forrestguice.suntimeswidget.calculator.SuntimesRiseSetDataset;
+import com.forrestguice.suntimeswidget.settings.WidgetSettings;
+
+import java.util.ArrayList;
+import java.util.Collections;
 
 public class LightMapDialog extends DialogFragment
 {
+    private static SuntimesUtils utils = new SuntimesUtils();
+
     private LightMapView lightmap;
-    private ImageView ic_night, ic_astro, ic_nautical, ic_civil, ic_day;
+    private LightMapKey field_night, field_astro, field_nautical, field_civil, field_day;
+    private int colorNight, colorAstro, colorNautical, colorCivil, colorDay;
+    private boolean showSeconds = true;
 
     private SuntimesRiseSetDataset data;
     public void setData( SuntimesRiseSetDataset data )
@@ -67,27 +77,20 @@ public class LightMapDialog extends DialogFragment
             Log.d("DEBUG", "LightMapDialog onCreate (restoreState)");
         }
 
-        dialog.setOnShowListener(new DialogInterface.OnShowListener()
-        {
-            @Override
-            public void onShow(DialogInterface dialogInterface)
-            {
-                lightmap.updateViews(data);
-            }
-        });
-
         themeViews(dialog.getContext());
+        updateViews();
         return dialog;
     }
 
     public void initViews(View dialogView)
     {
         lightmap = (LightMapView)dialogView.findViewById(R.id.info_time_lightmap);
-        ic_night = (ImageView)dialogView.findViewById(R.id.info_time_lightmap_key_night_icon);
-        ic_astro = (ImageView)dialogView.findViewById(R.id.info_time_lightmap_key_astro_icon);
-        ic_nautical = (ImageView)dialogView.findViewById(R.id.info_time_lightmap_key_nautical_icon);
-        ic_civil = (ImageView)dialogView.findViewById(R.id.info_time_lightmap_key_civil_icon);
-        ic_day = (ImageView)dialogView.findViewById(R.id.info_time_lightmap_key_day_icon);
+
+        field_night = new LightMapKey(dialogView, R.id.info_time_lightmap_key_night_icon, R.id.info_time_lightmap_key_night_label, R.id.info_time_lightmap_key_night_duration);
+        field_astro = new LightMapKey(dialogView, R.id.info_time_lightmap_key_astro_icon, R.id.info_time_lightmap_key_astro_label, R.id.info_time_lightmap_key_astro_duration);
+        field_nautical = new LightMapKey(dialogView, R.id.info_time_lightmap_key_nautical_icon, R.id.info_time_lightmap_key_nautical_label, R.id.info_time_lightmap_key_nautical_duration);
+        field_civil = new LightMapKey(dialogView, R.id.info_time_lightmap_key_civil_icon, R.id.info_time_lightmap_key_civil_label, R.id.info_time_lightmap_key_civil_duration);
+        field_day = new LightMapKey(dialogView, R.id.info_time_lightmap_key_day_icon, R.id.info_time_lightmap_key_day_label, R.id.info_time_lightmap_key_day_duration);
     }
 
     @SuppressWarnings("ResourceType")
@@ -101,23 +104,23 @@ public class LightMapDialog extends DialogFragment
         };
         TypedArray typedArray = context.obtainStyledAttributes(colorAttrs);
         int def = R.color.transparent;
-        int colorNight = ContextCompat.getColor(context, typedArray.getResourceId(0, def));
-        int colorAstro = ContextCompat.getColor(context, typedArray.getResourceId(1, def));
-        int colorNautical = ContextCompat.getColor(context, typedArray.getResourceId(2, def));
-        int colorCivil = ContextCompat.getColor(context, typedArray.getResourceId(3, def));
-        int colorDay = ContextCompat.getColor(context, typedArray.getResourceId(4, def));
+        colorNight = ContextCompat.getColor(context, typedArray.getResourceId(0, def));
+        colorAstro = ContextCompat.getColor(context, typedArray.getResourceId(1, def));
+        colorNautical = ContextCompat.getColor(context, typedArray.getResourceId(2, def));
+        colorCivil = ContextCompat.getColor(context, typedArray.getResourceId(3, def));
+        colorDay = ContextCompat.getColor(context, typedArray.getResourceId(4, def));
         typedArray.recycle();
 
-        colorizeImageView(ic_night, colorNight);
-        colorizeImageView(ic_astro, colorAstro);
-        colorizeImageView(ic_nautical, colorNautical);
-        colorizeImageView(ic_civil, colorCivil);
-        colorizeImageView(ic_day, colorDay);
+        colorizeImageView(field_night.icon, colorNight);
+        colorizeImageView(field_astro.icon, colorAstro);
+        colorizeImageView(field_nautical.icon, colorNautical);
+        colorizeImageView(field_civil.icon, colorCivil);
+        colorizeImageView(field_day.icon, colorDay);
     }
 
     private void colorizeImageView(ImageView view, int color)
     {
-        if (view.getBackground() != null)
+        if (view != null && view.getBackground() != null)
         {
             GradientDrawable d = (GradientDrawable) view.getBackground().mutate();
             d.setColor(color);
@@ -125,12 +128,159 @@ public class LightMapDialog extends DialogFragment
         }
     }
 
-    public void updateViews( SuntimesRiseSetDataset data )
+    public void updateViews()
+    {
+        if (data != null)
+            updateViews(data);
+    }
+
+    protected void updateViews( @NonNull SuntimesRiseSetDataset data )
     {
         if (lightmap != null)
         {
+            Context context = getContext();
+            field_civil.updateInfo(context, createInfoArray(data.civilTwilightLength()));
+            field_nautical.updateInfo(context, createInfoArray(data.nauticalTwilightLength()));
+            field_astro.updateInfo(context, createInfoArray(data.astroTwilightLength()));
+            field_night.updateInfo(context, createInfoArray(new long[] {data.nightLength()}));
+
+            long dayDelta = data.dayLengthOther() - data.dayLength();
+            field_day.updateInfo(context, createInfoArray(data.dayLength(), dayDelta, colorDay));
+
             lightmap.updateViews(data);
             Log.d("DEBUG", "LightMapDialog updated");
         }
     }
+
+    /**
+     * LightMapKey
+     */
+    private class LightMapKey
+    {
+        protected ImageView icon;
+        protected TextView label;
+        protected TextView text;
+
+        public LightMapKey(ImageView icon, TextView label, TextView duration)
+        {
+            this.icon = icon;
+            this.label = label;
+            this.text = duration;
+        }
+
+        public LightMapKey(@NonNull View parent, int iconRes, int labelRes, int durationRes)
+        {
+            icon = (ImageView)parent.findViewById(iconRes);
+            label = (TextView)parent.findViewById(labelRes);
+            text = (TextView)parent.findViewById(durationRes);
+        }
+
+        public void setVisible(boolean visible)
+        {
+            int visibility = (visible ? View.VISIBLE : View.GONE);
+            if (label != null) {
+                label.setVisibility(visibility);
+            }
+            if (text != null) {
+                text.setVisibility(visibility);
+            }
+            if (icon != null) {
+                icon.setVisibility(visibility);
+            }
+        }
+
+        public void updateInfo(Context context, LightMapKeyInfo[] info)
+        {
+            if (text == null || info == null || context == null)
+                return;
+
+            if (info.length == 1)
+            {
+                String duration = info[0].durationString(showSeconds);
+                if (info[0].delta > 0) {
+                    String s = context.getString(R.string.length_twilight1e_pos, duration, info[0].deltaString(showSeconds));
+                    if (info[0].durationColor != null)
+                        text.setText(SuntimesUtils.createColorSpan(s, duration, info[0].durationColor));
+                    else text.setText(new SpannableString(s));
+
+                } else if (info[0].delta < 0) {
+                    String s = context.getString(R.string.length_twilight1e_neg, duration, info[0].deltaString(showSeconds));
+                    if (info[0].durationColor != null)
+                        text.setText(SuntimesUtils.createColorSpan(s, duration, info[0].durationColor));
+                    else text.setText(new SpannableString(s));
+
+                } else {
+                    String s = context.getString(R.string.length_twilight1, duration);
+                    if (info[0].durationColor != null)
+                        text.setText(SuntimesUtils.createColorSpan(s, duration, info[0].durationColor));
+                    else text.setText(new SpannableString(s));
+                }
+                setVisible(true);
+
+            } else if (info.length >= 2) {
+                String s = context.getString(R.string.length_twilight2, info[0].durationString(showSeconds), info[1].durationString(showSeconds));
+                String delimiter = context.getString(R.string.length_delimiter);
+                text.setText(SuntimesUtils.createBoldColorSpan(s, delimiter, colorDay));
+                setVisible(true);
+
+            } else {
+                text.setText(new SpannableString(""));
+                setVisible(false);
+            }
+        }
+    }
+
+    /**
+     * LightMapKeyInfo
+     */
+    public static class LightMapKeyInfo
+    {
+        public LightMapKeyInfo(long duration, long delta)
+        {
+            this.duration = duration;
+            this.delta = delta;
+        }
+
+        public long duration = 0;
+        public Integer durationColor = null;
+        public String durationString(boolean showSeconds)
+        {
+            return utils.timeDeltaLongDisplayString(duration, showSeconds).toString();
+        }
+
+        public long delta = 0;
+        public Integer deltaColor = null;
+        public String deltaString(boolean showSeconds)
+        {
+            return utils.timeDeltaLongDisplayString(delta, showSeconds).toString();
+        }
+    }
+
+    public static LightMapKeyInfo[] createInfoArray(long durations, long delta, int color)
+    {
+        if (durations != 0)
+        {
+            LightMapKeyInfo[] info = new LightMapKeyInfo[1];
+            info[0] = new LightMapKeyInfo(durations, delta);
+            info[0].durationColor = color;
+            return info;
+
+        } else {
+            return new LightMapKeyInfo[0];
+        }
+    }
+
+    public static LightMapKeyInfo[] createInfoArray(long[] durations)
+    {
+        ArrayList<LightMapKeyInfo> info = new ArrayList<>();
+        for (int i=0; i<durations.length; i++)
+        {
+            if (durations[i] != 0)
+            {
+                info.add(new LightMapKeyInfo(durations[i], 0));
+            }
+        }
+        return info.toArray(new LightMapKeyInfo[0]);
+    }
+
 }