+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### ~
 
+* adds twilight durations to the lightmap dialog.
 * adds option to "places settings" to generate a list of world cities (by scanning locale defaults).
 * adds "show weeks" option to app and solstice widget (#153).
 * adds "show blue hour" option to app (#127).